+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Added
+
+- The `autoFocus` prop to focus on the input when the component mounts.
+
 ### Changed
 
 - No longer rely on `fetch` being available.